+++ lib/lightning/config.ex
@@ -251,6 +251,15 @@ defmodule Lightning.Config do
     defp promex_config do
       Application.get_env(:lightning, Lightning.PromEx, [])
     end
+
+    @impl true
+    def ui_metrics_tracking_enabled? do
+      Keyword.get(ui_metrics_tracking_config(), :enabled)
+    end
+
+    defp ui_metrics_tracking_config do
+      Application.get_env(:lightning, :ui_metrics_tracking, [])
+    end
   end
 
   @callback apollo(key :: atom() | nil) :: map()
@@ -281,6 +290,7 @@ defmodule Lightning.Config do
   @callback storage() :: term()
   @callback storage(key :: atom()) :: term()
   @callback token_signer() :: Joken.Signer.t()
+  @callback ui_metrics_tracking_enabled?() :: boolean()
   @callback usage_tracking() :: Keyword.t()
   @callback usage_tracking_cleartext_uuids_enabled?() :: boolean()
   @callback usage_tracking_cron_opts() :: [Oban.Plugins.Cron.cron_input()]
@@ -458,6 +468,10 @@ defmodule Lightning.Config do
 
   def promex_metrics_endpoint_token do
     impl().promex_metrics_endpoint_token()
+  end
+
+  def ui_metrics_tracking_enabled? do
+    impl().ui_metrics_tracking_enabled?()
   end
 
   defp impl do