+++ CHANGELOG.md
@@ -31,6 +31,8 @@ and this project adheres to
 - Indexes to foreign keys on `workflow_edges` and `steps` tables to try and
   alleviate slow loading of the job editor.
   [#2617](https://github.com/OpenFn/lightning/issues/2617)
+- Add temporary events to allow Lightning to log metrics reported by editors.
+  [#2617](https://github.com/OpenFn/lightning/issues/2617)
 
 ### Changed